backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
requests>=2.26.0
openai>=1.0.0
pinecone-client>=5.0.0
sentence-transformers>=2.2.2
numpy>=1.24.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.1.3
tiktoken>=0.7.0